requirements.txt
requests>=2.25.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
rasterio>=1.2.0
scipy>=1.7.0
geopy>=2.2.0
